+++ crates/web-client/test/address.test.ts
@@ -33,6 +33,7 @@ const instanceAddress = async ({
 const instanceNewAddressBech32 = async (page: Page, networkId: string) => {
   return await page.evaluate(async (bech32Prefix) => {
     const client = window.client;
+    const parsedNetworkId = window.helpers.parseNetworkId(bech32Prefix);
     const newAccount = await client.newWallet(
       window.AccountStorageMode.private(),
       true
@@ -41,7 +42,7 @@ const instanceNewAddressBech32 = async (page: Page, networkId: string) => {
       newAccount.id(),
       "BasicWallet"
     );
-    return address.toBech32(bech32Prefix);
+    return address.toBech32(parsedNetworkId);
   }, networkId);
 };
 
@@ -50,8 +51,9 @@ const instanceAddressFromBech32 = async (
   bech32EncodedAddress: string
 ) => {
   return await page.evaluate(async (bech32EncodedAddress) => {
+    const parsedNetworkId = window.helpers.parseNetworkId("mtst");
     const address = window.Address.fromBech32(bech32EncodedAddress);
-    return address.toBech32("mtst") === bech32EncodedAddress;
+    return address.toBech32(parsedNetworkId) === bech32EncodedAddress;
   }, bech32EncodedAddress);
 };